LPS-199164 Inline XmlRpcParser into XmlRpcUtil
# breaking_change_report

## What portal-kernel/src/com/liferay/portal/kernel.xmlrpc.XmlRpcParser.java

XmlRpcParser is inlined into XmlRpcUtil

## Why

XmlRpcParser and XmlRpcUtil both are tools to work with XmlRpc and depend on each other. Merging them unifies working with XmlRpc.

## Alternatives

Use the same methods in XmlRpcUtil instead
